+++ api/providers/generics/__init__.py
@@ -1,0 +1,3 @@
+"""
+This module contains the abstract definitions of the methods a provider is required to provide.
+"""
