+++ api/providers/simple_provider/provider.py
@@ -2,7 +2,9 @@ import os
 
 from providers.simple_provider.registration import FileRegistration
 from modules.path_constants import PATHS
-from utilities.GenericProvider import GenericProvider, GenericClub, GenericRegistration
+from providers.generics.GenericProvider import GenericProvider
+from providers.generics.GenericClub import GenericClub
+from providers.generics.GenericRegistration import GenericRegistration
 from utilities.exception_router import NotFound, Conflict
 
 
@@ -16,7 +18,7 @@ class ClubFaker(GenericClub):
     This provider doesn't allow for club-based registration.
     """
 
-    def get_registration(self, card_id) -> GenericRegistration:
+    def registration(self, card_id) -> GenericRegistration:
         fpath = os.path.join(PATHS.OUTPUT, "{}_{}".format(self.name, card_id))
         if os.path.isfile(fpath):
             return FileRegistration(self.name, card_id)
